+++ cherab/tools/inversions/nnls.py
@@ -21,13 +21,16 @@ import numpy as np
 import scipy
 
 
-def invert_regularised_nnls(w_matrix, b_vector, alpha=0.01, tikhonov_matrix=None):
-    """
+def invert_regularised_nnls(w_matrix, b_vector, alpha=0.01, tikhonov_matrix=None, **kwargs):
+    r"""
     Solves :math:`\mathbf{b} = \mathbf{W} \mathbf{x}` for the vector :math:`\mathbf{x}`,
     using Tikhonov regulariastion.
 
     This is a thin wrapper around scipy.optimize.nnls, which modifies
     the arguments to include the supplied Tikhonov regularisation matrix.
+
+    The values of w_matrix, b_vector and alpha * tikhonov_matrix are notmalised
+    by max(b_vector) before passing them to scipy.optimize.nnls().
 
     :param np.ndarray w_matrix: The sensitivity matrix describing the coupling between the
       detectors and the voxels. Must be an array with shape :math:`(N_d, N_s)`.
@@ -36,6 +39,7 @@ def invert_regularised_nnls(w_matrix, b_vector, alpha=0.01, tikhonov_matrix=None
       the regularisation strength of the tikhonov matrix.
     :param np.ndarray tikhonov_matrix: The tikhonov regularisation matrix operator, an array
       with shape :math:`(N_s, N_s)`. If None, the identity matrix is used.
+    :param **kwargs: Keyword arguments passed to scipy.optimize.nnls.
     :return: (x, norm), the solution vector and the residual norm.
 
     .. code-block:: pycon
@@ -60,6 +64,9 @@ def invert_regularised_nnls(w_matrix, b_vector, alpha=0.01, tikhonov_matrix=None
     d_vector = np.zeros(m+n)
     d_vector[0:m] = b_vector[:]
 
-    x_vector, rnorm = scipy.optimize.nnls(c_matrix, d_vector)
+    # Normalise c_matrix and d_vector to avoid possible issues with the nnls termination criteria.
+    vmax = d_vector.max()
 
-    return x_vector, rnorm
+    x_vector, rnorm = scipy.optimize.nnls(c_matrix / vmax, d_vector / vmax, **kwargs)
+
+    return x_vector, rnorm * vmax